+++ LookupAnything/Components/LookupMenu.cs
@@ -16,7 +16,7 @@ using StardewValley.Menus;
 namespace Pathoschild.Stardew.LookupAnything.Components
 {
     /// <summary>A UI which shows information about an item.</summary>
-    internal class LookupMenu : BaseMenu, IDisposable
+    internal class LookupMenu : BaseMenu, IScrollableMenu, IDisposable
     {
         /*********
         ** Fields
@@ -228,15 +228,13 @@ namespace Pathoschild.Stardew.LookupAnything.Components
             this.ExitOnNextTick = true;
         }
 
-        /// <summary>Scroll up the menu content by the specified amount (if possible).</summary>
-        /// <param name="amount">The amount to scroll, or <c>null</c> to scroll by <see cref="ScrollAmount"/>.</param>
+        /// <inheritdoc />
         public void ScrollUp(int? amount = null)
         {
             this.CurrentScroll -= amount ?? this.ScrollAmount;
         }
 
-        /// <summary>Scroll down the menu content by the specified amount (if possible).</summary>
-        /// <param name="amount">The amount to scroll, or <c>null</c> to scroll by <see cref="ScrollAmount"/>.</param>
+        /// <inheritdoc />
         public void ScrollDown(int? amount = null)
         {
             this.CurrentScroll += amount ?? this.ScrollAmount;